backend/requirements.txt
# Framework Web
fastapi[standard]
uvicorn
python-dotenv
pydantic

# Ecosistema LangChain
langchain
langchain-core
langchain-community
langchain-groq
langchain-google-genai
langchain-text-splitters
langchain-qdrant
langgraph

# ORM y Base de datos relacional
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings

# Vector DB y Búsqueda
qdrant-client
duckduckgo-search

# Procesamiento de documentos de pdfs y web scraping
pypdf
trafilatura

